kiemtra01/staff-service/requirements.txt
Django==5.1.7
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
django-cors-headers==4.6.0
PyMySQL==1.1.1
cryptography==44.0.2
requests==2.32.3
